Add scandir to shutil.
