+++ tests/test_smbclient_os.py
@@ -1285,16 +1285,6 @@ def test_scamdir_with_pattern(smb_share):
     assert names == ["file-test1.txt"]
 
 
-def test_scandir_local():
-    assert not is_remote_path(os.path.dirname(__file__))
-    assert not os.path.dirname(__file__).startswith(u"//localhost/share-encrypted/Pýtæs†-")
-    path = os.path.dirname(__file__)
-    assert not is_remote_path(path) and not path.startswith(u"//localhost/share-encrypted/Pýtæs†-")
-    scanner = smbclient.scandir(path)
-    contents = list(scanner)
-    assert any(os.path.basename(__file__) == entry.name for entry in contents)
-
-
 @pytest.mark.skipif(os.name != "nt" and not os.environ.get('SMB_FORCE', False),
                     reason="cannot create symlinks on Samba")
 def test_scandir_with_symlink(smb_share):
